Refactor: dto 수정 및 업데이트시 뷰에서 로고사진 업데이트
