+++ src/main/java/shop/mtcoding/job/dto/recruitmentPost/RecruitmentPostRespDto.java
@@ -7,11 +7,37 @@ import lombok.Setter;
 
 public class RecruitmentPostRespDto {
 
-    @Getter
     @Setter
-    public static class PostRespDto {
+    @Getter
+    public static class RecruitmentPostDetailRespDto {
         private Integer id;
+        private String enterpriseId;
         private String title;
+        private String career;
+        private String education;
+        private String pay;
+        private String sector;
+        private String position;
+        private String content;
+        private String address;
+        private String enterpriseName;
+        private Timestamp createdAt;
+    }
+
+    @Setter
+    @Getter
+    public static class RecruitmentPostListRespDto {
+        private Integer id;
+        private String enterpriseId;
+        private String title;
+        private String career;
+        private String education;
+        private String pay;
+        private String sector;
+        private String position;
+        private String content;
+        private String address;
+        private String enterpriseLogo;
         private String enterpriseName;
         private Timestamp createdAt;
     }
